Refactor account creation for owners. Document code.
I have documented the different functions used in web3.service.js in order to make the code more readable. Also, I have changed the way
the owners accounts are being created in order to allow them to sign transactions, i.e. be able to run the needed operations inside of
the blockchain. There is still improvement to be done in this part.

Next steps would be:

* Add ether to the owners accounts such that they will be able to execute the transactions. This should be done inside `createAccounts`
function.
* Finish the implementation of the `createDeliveryNote` function. Up to this point, there is a base implementation in which we are able
to register the devices in the Blockchain, but then we need to be able to add these devices into the DeliveryNote and send it from
OwnerA to OwnerB.
* Start with the DeliveryNote reception by OwnerB and the subsequent ownership transfer from OwnerA.